+++ front/app/components/UI/UserName/index.tsx
@@ -15,7 +15,7 @@ import { IUserData } from 'services/users';
 
 // i18n
 import { injectIntl } from 'utils/cl-intl';
-import { InjectedIntlProps } from 'react-intl';
+import { WrappedComponentProps } from 'react-intl';
 import messages from './messages';
 
 const Name = styled.span<{
@@ -75,7 +75,7 @@ interface Props extends StyleProps {
   hideLastName?: boolean;
 }
 
-const UserName = (props: Props & InjectedIntlProps) => {
+const UserName = (props: Props & WrappedComponentProps) => {
   const {
     intl: { formatMessage },
     userId,