+++ front/app/containers/CookiePolicy/index.tsx
@@ -3,9 +3,8 @@ import React from 'react';
 import { Helmet } from 'react-helmet';
 
 // i18n
+import { FormattedMessage, injectIntl, useIntl } from 'utils/cl-intl';
 import messages from './messages';
-import { FormattedMessage, injectIntl } from 'utils/cl-intl';
-import { InjectedIntlProps } from 'react-intl';
 
 // events
 import eventEmitter from 'utils/eventEmitter';
@@ -41,7 +40,8 @@ const StyledButton = styled.button`
   }
 `;
 
-const CookiePolicy = ({ intl: { formatMessage } }: InjectedIntlProps) => {
+const CookiePolicy = () => {
+  const { formatMessage } = useIntl();
   const openConsentManager = () => {
     eventEmitter.emit('openConsentManager');
   };