+++ front/app/modules/commercial/id_id_card_lookup/components/IdCardLookupForm.tsx
@@ -29,7 +29,7 @@ import useAuthUser from 'hooks/useAuthUser';
 import { verifyIDLookup } from '../services/verify';
 
 // i18n
-import { InjectedIntlProps } from 'react-intl';
+import { WrappedComponentProps } from 'react-intl';
 import { injectIntl, FormattedMessage } from 'utils/cl-intl';
 import messages from '../messages';
 import T from 'components/T';
@@ -46,7 +46,7 @@ interface Props {
   method: IDLookupMethod;
 }
 
-const VerificationFormLookup = memo<Props & InjectedIntlProps>(
+const VerificationFormLookup = memo<Props & WrappedComponentProps>(
   ({ onCancel, onVerified, showHeader, inModal, className, method, intl }) => {
     const authUser = useAuthUser();
 
@@ -190,4 +190,4 @@ const VerificationFormLookup = memo<Props & InjectedIntlProps>(
   }
 );
 
-export default injectIntl<Props>(VerificationFormLookup);
+export default injectIntl(VerificationFormLookup);