+++ front/app/services/formCustomFields.ts
@@ -146,6 +146,7 @@ export interface Result {
   question: Multiloc;
   totalResponses: number;
   answers: Answer[];
+  required: boolean;
 }
 
 export interface SurveyResultData {