+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 - [CL-1811] New Oostende verification method
 
+## 2022-10-20
+
+### Fixed
+
+- [CL-1814] Fixed bug when opening the verification model after having a verification error
+
 ## 2022-10-18
 
 ### Fixed